Moved CfgWeapons to its own file
